Flytter OppsummeringStønadsperioder til stønadsvilkår og flytter inn Box i komponenten då den ikke lengre brukes på tvers av flere komponenter
